requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic[email]
supabase==2.10.0
langchain==0.3.10
langchain-community==0.3.8
langchain-openai==0.2.10
openai==1.57.0
beautifulsoup4==4.12.3
requests==2.32.3
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
crawl4ai
pypdf2==3.0.1
pyMuPDF
unstructured[docx,xlsx]
python-docx==1.1.2
openpyxl==3.1.5
tiktoken==0.8.0
numpy<2,>=1.26.2
pandas==2.2.3
aiofiles==24.1.0
jinja2==3.1.4
pytest==8.3.4
pytest-asyncio==0.25.0
stripe==10.5.0
